[snarkyjs] get permissions working
